+++ custom/justech_l10n_do_ncf/tests/__init__.py
@@ -1,2 +1,3 @@
 from . import test_justech_l10n_do_ncf
 from . import test_ncf_sprint2
+from . import test_ncf_sprint2_part2