+++ custom/justech_l10n_do_ncf/validators/__init__.py
@@ -1,2 +1,3 @@
 """Validadores NCF específicos del módulo (reglas puras)."""
 from . import duplicate_scope
+from . import business_rules